Subindo projeto sem erros de maven e com minhas api junto, ainda não …

+++ dev/back-end/BackLaizeTech/BackLaizeTech/src/main/kotlin/laize_tech/back/repository/ProdutoRepository.kt
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Modifying
 import org.springframework.data.jpa.repository.Query
 import java.time.LocalDate
 
-interface ProdutoRepository : JpaRepository<Produto, Int> {
+interface ProdutoRepository : JpaRepository<Produto, Long> {
 
 //    @Transactional
 //    @Modifying
